+++ pennylane_lightning/core/src/measurements/MeasurementsBase.hpp
@@ -17,7 +17,6 @@
  */
 #pragma once
 
-#include <chrono>
 #include <random>
 #include <string>
 #include <vector>
@@ -81,10 +80,8 @@ template <class StateVectorT, class Derived> class MeasurementsBase {
      * @param seed Seed
      */
     void setRandomSeed() {
-        auto a = std::chrono::system_clock::now();
-        std::time_t b = std::chrono::system_clock::to_time_t(a);
-        // std::random_device rd;
-        setSeed(b);
+        std::random_device rd;
+        setSeed(rd());
     }
 
     /**